Transfer: completely abstract `DatbaseAdapter` from core export/import
... to allow multiple export versions.
